chore(tsconfig): change tsconfig to improve build
This adds declarations to the build which unfortunately also means that we need to disallow JS.
Also this changes the libs which are included and removes some redundancy.
The build also only includes components which are linked in the `index.js` file.
The build will exclude test and story files as they are not needed as a dependency.