feat(table): this introduces the table wrapping component
This introduces a Table component which expects data as props and Columns as children. Data is a
Array of any. It will use the props from the Column Component to render a proper table. This also
introduces the horizontal scrolling for the table if the content is bigger then the total width.
Although the first column always stays fixed.

close DCOS-38065, DCOS-37579