feat: Dropdownable
Dropdownable is a utility component that facilitates a simple dropdownable interaction: if the dropdown is open, a component prop is rendered relative to the position of the Dropdownable's children. When the user clicks outide of the component, it is hidden. The dropdown is rendered using an overlay that is appended to the document body.