refactor(Toaster): children as source-of-truth
- Refactored Toaster to always trust it's children as the source of truth
for which toasts should be displayed. This requires the app using
toaster to keep and accurate list of toasts and handle removing toasts
from that list whenever onDismiss callback is invoked.
- Updated `autodismiss` code to also ensure onDismiss callback is called.
- Updated Toaster to use toast.id for the Transition key to fix a bug
where new toasts would not transition in since they had the same key as
an older toast that has been dismissed

Closes DCOS-49609