refactor(breadcrumb): for readability
we now have an intersperse-function that communicates that we want to put
separators inbetween the elements. also we changed to using padding-left instead
of padding right to make the special case of NOT needing padding for the 1st
instead of the last element, which reduces mental overhead (at least for me).
compare:

- [padding("right", "xs")]: index !== crumbsArr.length - 1
+ [padding("left", "xs")]: i !== 0