feat(table): adds browser-native tooltip to all table cells
We want to ensure that table cell text content is always accessible to the user, even when a cell is truncated.
This commit gives every Table cell DOM node ]2;]1; attribute so users get a browser-native tooltip when hovering the cell.

This work was done to support DCOS-39095 (see Jira comments for better context)