refactor: don't rerender allTheCells allTheTime
Have a look at ContentCell in the Table-Component. Its sole purpose is to
highlight rows once a CheckboxTable has a disabled or checked row.

To do that it's computing css in a very dynamic way. That dynamic computation
breaks reacts reconciliation - effectively forcing every cell to rerender on
every update (a scroll-event is an update for the MultiGrid we use).

We *should* get rid of the dynamic nature of `ContentCell`. Unfortunately i
don't know how to do that, so this is rather a POC that breaks disabled and
checked rows in the `CheckboxTable`.