chore: allow a Toaster with a single child
this is rather not about this being a really common use case, but about
introducing some patterns that allow us to better work with types in the future.
we now have a place to add/override libraries types in order to work around
the usual type-shortcomings most libraries still have.

As a side effect of enhancing the type for React.Children.toArray some typecasts
are now obsolete and have been removed.