fix: adapt grid-template when changing columns
we formerly computed the grid-template out of the columns present in the table's
internal state, which saves every column ever known to that table.

we now introduce a separate type in order to enhance the Column-type with a
`width` for table-internal use. that makes it easy to properly set the
grid-template depending on the columns that currently shall be shown. Due to the
internal table state, it also remembers the widths for columns that are removed
and added again later.