feat(clicktocopybutton): adds a default tooltip when text is copied
BREAKING CHANGE:
If a user has implemented behavior where a ClickToCopyButton that shows a tooltip when the text is copied, two tooltips will be shown. The first tooltip will be the default tooltip text that was added in this commit. The second tooltip would be the one that was added when doing a custom tooltip implementation.

Closes D2IQ-71305