feat: add block display override to icon
Added a `block` flag to the Icon prop to allow changing the svg from the
default `display: inline` to `display: block` this will allow an icon to
use `margin: auto` inside flex. This is useful when displaying an icon
next to text and we want them to be aligned.

I did not change the default since it would break the rendering for
other components.

I initially set `block` as a default prop set to false, but this then
broke every snapshot. Instead I just made it an optional prop.

Alternative to a block flag we could of made a display prop the defaults
to inline and can be set to any valid display CSS value. I'm not opposed
to this but didn't need it now.