feat: slidetoggle
Added a new toggle component `SlideToggle` this is basically a checkbox
styled to be a slider with a check when checked.

I initially tried to implement this extending the ToggleComponent like
the CheckboxInput and RadioInputs do, but the styles were different
enough that I needed to make this its own component.

- Closes D2IQ-73430