fix: controlled FieldList row deletion
It seemed as though the onRemoveItem was never actually being called
based on the way that it was written, because the useEffect would not
fire, thus you'd end up with rendering issues. By tweaking a bit how
this function is called in the context, it appears to fix the issue.

I don't know why this callback function was written with the additional
closure, but I preserved it to maintain backwards compatibility.

Addresses D2IQ-77909