feat: improve grid features and alignment options
BREAKING CHANGE
Before this fix columnCount was required, new behavior includes defaults for when it is not set to become a fluid grid that sets the amount of rows and columns based on available space.
Before you could not center grid items, now the optional centerItems prop can be used.
Other features include a data-cy prop, and className for further customizations.

Closes D2IQ-81475