chore: upgrade downshift
not sure exactly how this test was meant to work prior to this but it
seems like the user behavior is what we want, it just skips over the
disabled item and allows you to select the first available.

Otherwise some type updates to get it to work. Could use some visual
regression testing on these components.

Closes D2IQ-88824