fix: revert module change (#774)
This steps back a complete shift from CommonJS to esm modules, and
instead makes only the file that imports tablemark use esm. Then, we use
a dynamic import statement inside the build script to import that file
as a workaround. Thus we're able to leave tablemark as is for now.

Closes D2IQ-92904