refactor: hashmap and corresponding components (#772)
Refactored hashmap from a class to a function. Tried to make some type
improvements and also to refactor some of the tests. Had trouble using
react-test-renderer create with one of the tests that used a mock
function, possibly because of children further down the dom tree, so I
skipped it.