fix: pagination as links bug, fullscreen props (#773)
* fix: pagination as links bug, fullscreen props

This began as a refactor of several components away from classes. While
doing so I was able to remove the Delegate library that was throwing
react errors because it was very out of date. Additonally I found a bug
with the way proptypes were defined for headerComponent in two
components, and also a bug with the pagination when trying to use the
pager buttons as links.

* refactor: allow any to preserve bwards compat