refactor: typeahead component and stories
This one was more complicated but converted Typeahead from a class
component to a functional component. All functionality should be
preserved, but I did find that the CustomMenuEmptyState story did not
work as expected because the empty state should be shown when there are
no items. Also made a decision to remove the filteredlist typeahead
because it didn't really demonstrate any functionality of the component
itself, just of how a wrapping component could control its state. It
seemed superfluous.