feat: optional max-width for FullscreenView header
If child content of FullscreenView is wrapped in a container, it would
be nice to be able to contain the header of this component also for
visual consistency. Opted to keep the padding as is and just account for
it in the max width so that consumers can pass in our design tokens as
is without having to do the math themselves to account for the padding
left and right.

Addresses D2IQ-92266