test: clean up toast test suite
The toast test suite was tightly coupled with Toaster. This may have
been a copy paste issue or something, but the snapshots didn't even
reflect the markup of the Toast component. I have removed the wrapping
of each of these in Toaster, and a couple of tests that were targeting
behavior of Toaster rather than Toast. I have refactored to use RTL
instead of enzyme in other cases.