feat: make appchrome props optional
Allows for someone to use just the header or sidebar without needing to
declare both. Also add children prop and deprecates mainContent because
its a lot easier to use with children, as it allows JSX to simply be
nested rather than part of a prop. Found these while trying to use this
component during innovation days.