fix: make var check ssr friendly
document and window are not defined in node environments, so trying to
use ui-kit with an isomorphic framework like next will fail saying
document/window is not defined. Adding a check that window is not
undefined before trying to operate on document seems to fix the issue.
There may be more places like this in ui-kit but this is all I've run up
against so far.