fix: fix broken alias transformer and update design token docs
When style-dictionary was upgraded from 2.x.x to 3.x.x the data structure of the
design token changed, which meant that the alias transformer was using the value
of the token rather than the object path in order to generate the alias.

With the transformer fixed, this commit also updates the design token docs.

Closes D2IQ-95943